backend/pyproject.toml
[project]
name = "profile-optimizer-backend"
version = "0.1.0"
description = "Backend for White Rabbit Profile Optimizer"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.109.0",
    "uvicorn[standard]>=0.27.0",
    "sqlalchemy>=2.0.25",
    "alembic>=1.13.1",
    "asyncpg>=0.29.0",
    "pgvector>=0.2.4",
    "pydantic>=2.6.0",
    "pydantic-settings>=2.1.0",
    "anthropic>=0.18.1",
    "python-multipart>=0.0.9",
    "httpx>=0.26.0",
    "beautifulsoup4>=4.12.3",
    "clerk-backend-api>=1.0.0", # Or appropriate Clerk SDK if available, otherwise we'll use httpx/jwks
    "pyjwt[crypto]>=2.8.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=9.0.0",
    "pytest-asyncio>=0.23.0",
    "ruff>=0.11.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.pytest.ini_options]
pythonpath = ["."]
testpaths = ["tests"]
asyncio_mode = "auto"

[tool.ruff]
target-version = "py311"

[tool.ruff.lint]
ignore = [
    "E712",  # SQLAlchemy requires == True/False for column comparisons
]

[tool.ruff.lint.per-file-ignores]
"app/main.py" = ["E402"]  # FastAPI app must be created before importing routers

[tool.hatch.build.targets.wheel]
packages = ["app"]
